Version 0.2.0
Throw away the mail-merge implementation and instead use a straightforward search/replace

This is because, while I could replace the merge-fields on-screen, Word keeps a reference to them somewhere and they reappear when you try to print the document.

The new implementation uses mustache style squiggly brackets to highlight fields.