Updated the merge document to deal with a docx where the merge field tags had been mixed up with Word formatting tags
Instead of a simple node containiner "{{ FIELD }}" this document has a tree of nodes, with a text element of "{{", then some formatting nodes, then "F" then more formatting then "IELD" then even more formatting followed by the "}}"

Instead of blindly expecting everything to be contained in a single node, the updated `tag_nodes` routine works by finding the first "{{".  Then it finds the ancestor node that contains both this opening "{{" node and the closing "}}" node.  It now has the entirety of the merge field tag, so it can safely extract the field name from it.