Bug fixes
When the docx has formatting attributes embedded in it, it can mangle the field tags.  However, the previous version, while it located the field tags and substituted them correctly, it did not reinsert them into the document properly.

Now, it searches for paragraph (w:p) tags instead of text (w:t) tags.  This means it no longer needs to walk the hierarchy to find the parent.

A paragraph may contain multiple field tags, so the field_names_from(tag_node) has been altered accordingly.

And when doing the substitution, it looks for paragraph formatting nodes (w:pPr) and run formatting nodes (w:rPr).  When performing the substitution, it replaces all the nodes underneath the paragraph.  Then inserts a paragraph properties node, then a run node which holds the run properties, then a w:t text node with the substituted text.

This is not perfect.  If the paragraph containing the field tags has individual formatting applied (for example, some words are in bold, some in a different colour) that formatting will be lost - and the same styles applied to the entire paragraph.

But it works and the output is visible in pretty much the same style as from the original template.

Adding in individual formatting is possible - search all w:r nodes for their w:rPr nodes, then rebuild it.  But if a field tag is split across multiple run nodes (because individual letters or combinations within the field tag have different formatting - or as is more likely, Word has repeated the same formatting across multiple run nodes representing the field tag - then rebuilding the substituted field tag will be tricky.

Not impossible but this works for now.