Basic framework for embedding images into docx files
This adds the image contents as separate streams into the output docx (word/media/whatever) and updates the relationships document (word/_rels/whatever.xml.rels) with relationships tying an ID to the stream.  And it adds an embedded image node into the document, so it can do the substitution.

However, the actual embedded image is not appearing in the final document yet!